requirements.txt
Flask==2.0.1
Flask-Babel==2.0.0
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.46
requests==2.26.0
python-dotenv==0.19.0
markdown==3.3.6
Babel==2.9.1
Werkzeug==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
itsdangerous==2.0.1
click==8.0.1
